Updated ffmpeg and fixed x64 builds.

+++ horseradish/src/engine/logger.cpp
@@ -177,7 +177,7 @@ namespace hr { namespace engine
 			return true;
 
 		bool hasFormattedText;
-		unsigned int entryDataSize;
+		size_t entryDataSize;
 		if (!Logger::checkEntryData(entryData, hasFormattedText, entryDataSize))
 			return false;
